Consolidated Vendor/FoodTruck models
Combined the Vendor and FoodTruck model classes.

Added configs for all other current model types and added properties to
the Context to make models available.

Added collection of Badges to FoodRUser. This will likely have to be
changed to a different object later. One that will join the Badge and
the User along with other details such as "Date Awarded".