Refactored AuthController
Instead of taking direct dependencies on Identity Framework objects
UserManager and RoleManager, the controller now takes a dependency on a
IUserService. The implementation UserService takes the Identity
Framework pieces as dependencies and gives us a single service to
consume from any controller that needs to get access to or update user
information.