Created FoodR.Data project
Moved data models and configuration to new project, refactored
namespaces and code a little bit for better organization. The purpose of
this is to allow use of the data model outside of the web app. We can
make console apps and test harnesses that use the data model and db
storage. This will help me in refactoring the data seeding.