Moved ComposeModelConfig to DbModelBuilder
The Extension Method ComposeModelConfig was originally written to hang
off of the DbContext but it never actually uses it. This leaves an
unused parameter on the extension method which can trigger a warning in
static code analysis. The method took another parameter that it does
take action upon, the DbModelBuilder, and so the extension method has
been moved to this type.